Moving load/save to a persistence module
