+++ pyCoilGen/pyCoilGen_release.py
@@ -20,7 +20,7 @@ from .helpers.visualisation import visualize_vertex_connections, visualize_compa
 from .helpers.timing import Timing
 
 # For saving Pickle files
-from .helpers.common import save
+from .helpers.persistence import save
 
 # From original project
 from .sub_functions.read_mesh import read_mesh
